+++ libnopegl/src/nodes.c
@@ -585,7 +585,7 @@ int ngl_node_param_swap_elem(struct ngl_node *node, const char *key,
     return ret;
 }
 
-static int node_invalidate_branch(struct ngl_node *node)
+int ngli_node_invalidate_branch(struct ngl_node *node)
 {
     node->last_update_time = -1;
     if (node->cls->invalidate) {
@@ -595,7 +595,7 @@ static int node_invalidate_branch(struct ngl_node *node)
     }
     struct ngl_node **parents = ngli_darray_data(&node->parents);
     for (size_t i = 0; i < ngli_darray_count(&node->parents); i++) {
-        int ret = node_invalidate_branch(parents[i]);
+        int ret = ngli_node_invalidate_branch(parents[i]);
         if (ret < 0)
             return ret;
     }
@@ -638,7 +638,7 @@ static int node_param_update(struct ngl_node *node, const struct node_param *par
             return ret;
     }
 
-    return node_invalidate_branch(node);
+    return ngli_node_invalidate_branch(node);
 }
 
 #define FORWARD_TO_PARAM(type, ...)                                     \