+++ CHANGELOG.md
@@ -28,6 +28,7 @@ Versioning](https://semver.org/spec/v2.0.0.html) for `libnopegl`.
 - `ngl_node_param_swap_elem()` to allow users to swap/re-order elements within
   a list parameter
 - `Group.children` can be live re-ordered
+- `CustomTexture` node which allows implementing custom rendering logic
 
 ### Changed
 - `Text.font_files` text-based parameter is replaced with `Text.font_faces` node